@@ -14,67 +14,67 @@
     "build:aot": "ngc -p ./tsconfig-aot.json"
   },
   "dependencies": {
-    "@angular/animations": "~5.0.0",
-    "@angular/common": "~5.0.0",
-    "@angular/compiler": "~5.0.0",
-    "@angular/compiler-cli": "~5.0.0",
-    "@angular/core": "~5.0.0",
-    "@angular/forms": "~5.0.0",
-    "@angular/http": "~5.0.0",
-    "@angular/platform-browser": "~5.0.0",
-    "@angular/platform-browser-dynamic": "~5.0.0",
-    "@angular/platform-server": "~5.0.0",
-    "@angular/router": "~5.0.0",
-    "lodash": "~4.17.5",
-    "rxjs": "~5.5.2",
-    "zone.js": "~0.8.18"
+    "@angular/animations": "~8.1.2",
+    "@angular/common": "~8.1.2",
+    "@angular/compiler": "~8.1.2",
+    "@angular/compiler-cli": "~8.1.2",
+    "@angular/core": "~8.1.2",
+    "@angular/forms": "~8.1.2",
+    "@angular/platform-browser": "~8.1.2",
+    "@angular/platform-browser-dynamic": "~8.1.2",
+    "@angular/platform-server": "~8.1.2",
+    "@angular/router": "~8.1.2",
+    "lodash": "~4.17.14",
+    "rxjs": "~6.5.2",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@ngtools/webpack": "~1.8.0",
-    "@types/jasmine": "~2.6.2",
-    "@types/lodash": "~4.14.105",
-    "@types/node": "^8.0.47",
+    "@ngtools/webpack": "~8.1.1",
+    "@types/jasmine": "~3.3.13",
+    "@types/lodash": "~4.14.136",
+    "@types/node": "^12.6.2",
     "angular2-router-loader": "~0.3.5",
     "angular2-template-loader": "~0.6.2",
-    "awesome-typescript-loader": "~3.4.1",
+    "awesome-typescript-loader": "~5.2.1",
     "babel": "~6.23.0",
-    "babel-core": "~6.26.0",
-    "babel-loader": "~7.1.2",
+    "babel-core": "~6.26.3",
+    "babel-loader": "~8.0.6",
     "babel-preset-es2015": "~6.24.1",
-    "clean-webpack-plugin": "~0.1.17",
-    "css-loader": "~0.28.7",
-    "del": "~3.0.0",
-    "del-cli": "~1.1.0",
-    "es6-promise": "~4.1.1",
-    "es6-shim": "~0.35.3",
-    "extract-text-webpack-plugin": "~3.0.2",
-    "file-loader": "~1.1.5",
-    "html-loader": "~0.5.1",
-    "html-webpack-plugin": "~2.30.1",
-    "http-server": "~0.10.0",
-    "istanbul-instrumenter-loader": "~3.0.0",
-    "jasmine-core": "~2.8.0",
-    "karma": "~1.7.1",
+    "clean-webpack-plugin": "~3.0.0",
+    "css-loader": "~3.0.0",
+    "del": "~5.0.0",
+    "del-cli": "~2.0.0",
+    "es6-promise": "~4.2.8",
+    "es6-shim": "~0.35.5",
+    "mini-css-extract-plugin": "~0.7.0",
+    "file-loader": "~4.0.0",
+    "html-loader": "~0.5.5",
+    "html-webpack-plugin": "~3.2.0",
+    "http-server": "~0.11.1",
+    "istanbul-instrumenter-loader": "~3.0.1",
+    "jasmine-core": "~3.4.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage": "~1.1.1",
-    "karma-coverage-istanbul-reporter": "~1.3.0",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage": "~1.1.2",
+    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma-jasmine": "~2.0.1",
     "karma-phantomjs-launcher": "~1.0.4",
     "karma-sourcemap-loader": "~0.3.7",
-    "karma-webpack": "~2.0.5",
-    "node-sass": "~4.6.0",
-    "null-loader": "~0.1.1",
-    "optimize-css-assets-webpack-plugin": "~3.2.0",
-    "raw-loader": "~0.5.1",
-    "resolve-url-loader": "2.2.1",
-    "sass-loader": "~6.0.6",
-    "style-loader": "~0.19.0",
+    "karma-webpack": "~4.0.2",
+    "node-sass": "~4.12.0",
+    "null-loader": "~3.0.0",
+    "optimize-css-assets-webpack-plugin": "~5.0.3",
+    "raw-loader": "~3.0.0",
+    "resolve-url-loader": "3.1.0",
+    "sass-loader": "~7.1.0",
+    "style-loader": "~0.23.1",
     "to-string-loader": "1.1.5",
-    "typescript": "~2.6.1",
-    "url-loader": "0.6.2",
-    "webpack": "~3.8.1",
-    "webpack-dev-server": "~2.9.4",
-    "webpack-merge": "~4.1.1",
-    "webpack-stream": "~4.0.0"
+    "typescript": "~3.4.3",
+    "url-loader": "2.0.1",
+    "webpack": "~4.35.3",
+    "webpack-cli": "^3.3.5",
+    "webpack-dev-server": "~3.7.2",
+    "webpack-merge": "~4.2.1",
+    "webpack-stream": "~5.2.1"
   }
 }
