@@ -2,62 +2,57 @@
   "name": "cash-stash-server",
   "version": "0.0.1",
   "private": true,
-  "engines": {
-    "node": "8.0.0"
-  },
   "scripts": {
     "build": "tsc",
     "start": "tsc && node dist/app.js",
     "test": "gulp test:client"
   },
   "dependencies": {
-    "ajv": "4.11.6",
-    "bcryptjs": "^2.4.0",
-    "body-parser": "~1.15.2",
+    "ajv": "6.10.1",
+    "bcryptjs": "^2.4.3",
+    "body-parser": "~1.19.0",
     "caller": "^1.0.1",
-    "cookie-parser": "~1.4.3",
-    "core-js": "^2.4.1",
-    "debug": "~4.1.0",
-    "express": "~4.14.0",
-    "hbs": "~3.1.0",
-    "helmet": "3.5.0",
-    "jsonwebtoken": "^7.2.1",
-    "lodash": "^4.17.4",
-    "log4js": "^1.1.1",
-    "mongoose": "4.9.7",
-    "mongoose-unique-validator": "^1.0.5",
-    "morgan": "~1.6.1",
-    "reflect-metadata": "^0.1.9",
-    "serve-favicon": "~2.3.0",
-    "swagger-parser": "^4.1.0",
-    "swagger-ui-express": "^3.0.8"
+    "cookie-parser": "~1.4.4",
+    "core-js": "^3.1.4",
+    "debug": "~4.1.1",
+    "express": "~4.17.1",
+    "hbs": "~4.0.4",
+    "helmet": "3.18.0",
+    "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.14",
+    "log4js": "^4.5.1",
+    "mongoose": "5.6.4",
+    "mongoose-unique-validator": "^2.0.3",
+    "morgan": "~1.9.1",
+    "reflect-metadata": "^0.1.13",
+    "serve-favicon": "~2.5.0",
+    "swagger-parser": "^8.0.0",
+    "swagger-ui-express": "^4.0.7"
   },
   "devDependencies": {
-    "@types/body-parser": "^1.16.4",
-    "@types/cookie-parser": "^1.3.30",
-    "@types/crypto-js": "^3.1.33",
-    "@types/express": "^4.0.36",
-    "@types/jsonwebtoken": "^7.2.1",
-    "@types/log4js": "^0.0.32",
-    "@types/mongodb": "^2.2.7",
-    "@types/mongoose": "^4.7.18",
-    "@types/node": "^7.0.14",
-    "del": "^2.2.2",
-    "del-cli": "^0.2.0",
-    "gulp": "^3.9.1",
+    "@types/body-parser": "^1.17.0",
+    "@types/cookie-parser": "^1.4.1",
+    "@types/crypto-js": "^3.1.43",
+    "@types/express": "^4.17.0",
+    "@types/jsonwebtoken": "^8.3.2",
+    "@types/mongodb": "^3.1.28",
+    "@types/mongoose": "^5.5.8",
+    "@types/node": "^12.6.2",
+    "del": "^5.0.0",
+    "del-cli": "^2.0.0",
+    "gulp": "^4.0.2",
     "gulp-env": "^0.4.0",
     "gulp-jasmine-node": "^1.0.7",
-    "gulp-load-plugins": "^1.4.0",
-    "gulp-plumber": "^1.1.0",
-    "gulp-watch": "^4.3.11",
-    "install": "^0.8.9",
-    "jasmine-node": "^1.14.5",
-    "nodemon": "^1.11.0",
-    "open": "0.0.5",
-    "request": "^2.79.0",
-    "run-sequence": "^1.2.2",
-    "supertest": "^3.0.0",
-    "typescript": "^2.4.1",
-    "v8-profiler": "5.7.0"
+    "gulp-load-plugins": "^2.0.0",
+    "gulp-plumber": "^1.2.1",
+    "gulp-watch": "^5.0.1",
+    "install": "^0.13.0",
+    "jasmine-node": "^3.0.0",
+    "nodemon": "^1.19.1",
+    "open": "6.4.0",
+    "request": "^2.88.0",
+    "run-sequence": "^2.2.1",
+    "supertest": "^4.0.2",
+    "typescript": "^3.5.3"
   }
 }
