@@ -1,9 +1,9 @@
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
-import { RouterTestingModule } from "@angular/router/testing";
-import { APP_BASE_HREF } from "@angular/common";
-import { HttpModule } from "@angular/http";
-import { NavbarComponent } from "./navbar.component";
-import { AuthService } from "../../auth/auth-service/auth.service";
+import {async, ComponentFixture, TestBed} from "@angular/core/testing";
+import {RouterTestingModule} from "@angular/router/testing";
+import {APP_BASE_HREF} from "@angular/common";
+import {HttpClientModule} from "@angular/common/http";
+import {NavbarComponent} from "./navbar.component";
+import {AuthService} from "../../auth/auth-service/auth.service";
 
 describe("NavbarComponent", () => {
 
@@ -23,7 +23,7 @@ describe("NavbarComponent", () => {
     TestBed.configureTestingModule({
       imports: [
         RouterTestingModule,
-        HttpModule
+        HttpClientModule
       ],
       declarations: [NavbarComponent],
       providers: [
