@@ -1,9 +1,10 @@
-import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
-import { Account } from '../account.model';
-import { HttpClient } from '@angular/common/http';
-import { FlashService } from '../../components/flash/flash.service';
-import { WraithRestDatasource } from '../../components/wraith-datasource/wraith.rest.datasource';
+import {Injectable} from '@angular/core';
+import {Observable} from 'rxjs';
+import {Account} from '../account.model';
+import {HttpClient} from '@angular/common/http';
+import {FlashService} from '../../components/flash/flash.service';
+import {WraithRestDatasource} from '../../components/wraith-datasource/wraith.rest.datasource';
+import {catchError} from "rxjs/operators";
 
 @Injectable()
 export class AccountService extends WraithRestDatasource<Account> {
@@ -14,41 +15,41 @@ export class AccountService extends WraithRestDatasource<Account> {
 
   getAccounts() {
     return this.getAll()
-      .catch((error) => {
+      .pipe(catchError((error) => {
         this.flashService.error(error);
         return Observable.throw(error)
-      });
+      }));
   };
 
   getAccount(id: string) {
     return this.get(id)
-      .catch((error) => {
+      .pipe(catchError((error) => {
         this.flashService.error(error);
         return Observable.throw(error)
-      });
+      }));
   };
 
   saveAccount(account: Account) {
     return this.insert(account)
-      .catch((error) => {
+      .pipe(catchError((error) => {
         this.flashService.error(error);
         return Observable.throw(error);
-      });
+      }));
   };
 
   updateAccount(id: string, account: Account) {
     return this.update(id, account)
-      .catch((error) => {
+      .pipe(catchError((error) => {
         this.flashService.error(error);
         return Observable.throw(error)
-      });
+      }));
   }
 
   deleteAccount(id: string) {
     return this.remove(id)
-      .catch((error) => {
+      .pipe(catchError((error) => {
         this.flashService.error(error);
         return Observable.throw(error)
-      });
+      }));
   }
 }
