@@ -1,9 +1,10 @@
-import { Injectable } from '@angular/core';
-import { User } from '../user.model';
-import { Observable } from 'rxjs';
-import { HttpClient, HttpErrorResponse } from '@angular/common/http';
-import { FlashService } from '../../components/flash/flash.service';
-import { WraithBaseRequest } from '../../components/wraith-datasource/wraith.base-request';
+import {Injectable} from '@angular/core';
+import {User} from '../user.model';
+import {Observable} from 'rxjs';
+import {catchError, map} from 'rxjs/operators';
+import {HttpClient, HttpErrorResponse} from '@angular/common/http';
+import {FlashService} from '../../components/flash/flash.service';
+import {WraithBaseRequest} from '../../components/wraith-datasource/wraith.base-request';
 
 @Injectable()
 export class AuthService extends WraithBaseRequest {
@@ -14,19 +15,19 @@ export class AuthService extends WraithBaseRequest {
 
   signup(user: User): Observable<any> {
     const body = JSON.stringify(user);
-    return this.httpClient.post(this.baseUrl + '/auth/signup', body, { headers: super.getHeader() })
-      .map((data: any) => {
+    return this.httpClient.post(this.baseUrl + '/auth/signup', body, {headers: super.getHeader()})
+      .pipe(map((data: any) => {
         localStorage.setItem('token', data.result[0].token);
         return true;
-      }).catch((error: HttpErrorResponse) => {
+      }), catchError((error: HttpErrorResponse) => {
         this.flashService.error(error.name + ' (' + error.message + ')');
         return Observable.throw(error)
-      });
+      }));
   };
 
   signin = (user: User) => {
     const body = JSON.stringify(user);
-    return this.httpClient.post(this.baseUrl + '/auth/signin', body, { headers: super.getHeader() })
+    return this.httpClient.post(this.baseUrl + '/auth/signin', body, {headers: super.getHeader()})
       .subscribe((data: any) => {
           localStorage.setItem('token', data.result[0].token);
           return true;
