@@ -41,9 +41,10 @@ export class Environment {
     this.logging = process.env.LOGGING_LEVEL || 'ERROR';
     // this.logging = process.env.LOGGING_LEVEL || 'TRACE';
     this.root = path.normalize(__dirname + '/../../..');
-    this.port = process.env.PORT || 3000;
-    // this.ipAddress = process.env.IP || '192.168.99.100';
-    this.ipAddress = process.env.IP || 'localhost';
+    this.port = parseInt(process.env.PORT) || 3000;
+    // this.port = 3000;
+    this.ipAddress = process.env.IP || '192.168.0.213';
+    // this.ipAddress = process.env.IP || 'localhost';
     this.secret = process.env.SECRET || 'secret';
     this.databaseName = process.env.databaseName || 'cash-stash';
     // this.databaseName = process.env.databaseName || 'cash-stash-development';
